separated out lib code
